LED Sync with timer
All LED turn on with timer.
It’s added sync and reset.
It must add hart only to sync with PC.
But it is send side only,so should to make get side by tomorrow.
Shuld to care thing to be uart program is,remove the while(1); from
last of main() function.